modules/zantara-infrastructure/light-agent/requirements.txt
flask==2.3.3
google-cloud-firestore==2.12.0
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.10.0
gunicorn==21.2.0
requests==2.31.0